--- src/InvalidUserInput.java
@@ -1,2 +1,0 @@
-public class InvalidUserInput extends Throwable {
-}
