--- src/MainView.java
@@ -1,10 +1,0 @@
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-
-public class MainView {
-    GugudanController gugudanController = new GugudanController();
-    public void start() throws IOException {
-    }
-
-}
