+++ src/controller/GugudanController.java
@@ -1,3 +1,7 @@
+package controller;
+
+import service.GugudanService;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
@@ -5,12 +9,13 @@ import java.io.InputStreamReader;
 public class GugudanController {
     private GugudanService gs = new GugudanService();
 
+    // TODO -> 서비스 로직과 컨트롤러 로직을 분리!~
+    public void run() throws IOException {
+        gs.run();
+    }
+
     // private GugudanBinaryService gbs = new GugudanBinaryService();
     // private GugudanSomeService gbs = new GugudanSomeService();
-
-    public GugudanController(GugudanView view) {
-        this.view = view;
-    }
 
     public void runBinaryGugudan() {
         // some code ...
@@ -20,10 +25,7 @@ public class GugudanController {
         // ... some code
     }
 
-    // TODO -> 서비스 로직과 컨트롤러 로직을 분리!~
-    public void run() throws IOException {
-        gs.run();
-    }
+
 
 
 }
