+++ src/exception/InvalidUserInput.java
@@ -1,0 +1,10 @@
+package exception;
+
+public class InvalidUserInput extends Throwable {
+    public InvalidUserInput() {
+    }
+
+    public InvalidUserInput(String message) {
+        super(message);
+    }
+}
